requirements.txt
streamlit
openai
fpdf2
python-dotenv
reportlab
flask



